+++ src/main/java/com/ganaseguros/apimovilweb/utils/constantes/ConstEstado.java
@@ -1,5 +1,5 @@
 package com.ganaseguros.apimovilweb.utils.constantes;
 
 public class ConstEstado {
-    public static final Integer ACTIVO = 1000;
+    public static final Long ACTIVO = 1000L;
 }
